protect delete route

+++ controllers/stuff.js
@@ -16,8 +16,22 @@ exports.modifyThing = (req, res, next) => {
 }
 
 exports.deleteThing = (req, res, next) => {
-    Thing.deleteOne({_id: req.params.id})
-        .then(() => res.status(200).json({message: 'Object Deleted'}))
+    Thing.findOne({_id: req.params.id})
+        .then((thing) => {
+            if (!thing) {
+                return res.status(404).json({
+                    error: new Error('Object not found')
+                });
+            }
+            if (thing.userId != req.auth.userId) {
+                return res.status(401).json({
+                    error: new Error('Request not authorized')
+                });
+            }
+            Thing.deleteOne({_id: req.params.id})
+                .then(() => res.status(200).json({message: 'Object Deleted'}))
+                .catch(error => res.status(400).json({error}));
+        })
         .catch(error => res.status(400).json({error}));
 }
 exports.getOneThing = (req, res, next) => {
